Removed repository dependency injection from controller. Implemented new service to work with repository. Wrapped controller methods with ResponseEntity.
