+++ src/main/java/com/example/coursework/service/NodeService.java
@@ -1,19 +1,16 @@
 package com.example.coursework.service;
 
-import com.example.coursework.model.State;
+import com.example.coursework.model.Node;
 import com.example.coursework.repository.NodeRepository;
-import org.springframework.context.annotation.Scope;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.RequestBody;
 
-import java.util.Random;
+import java.util.List;
 
 @Service
-@Scope("singleton")
 public class NodeService
 {
-    private double temperature = 600;
-    private double minTemperature = 10;
-    private final double temperatureConst = 0.5;
     private final NodeRepository nodeRepository;
 
     public NodeService(NodeRepository nodeRepository)
@@ -21,48 +18,13 @@ public class NodeService
         this.nodeRepository = nodeRepository;
     }
 
-    public State startSimulatedAnnealing()
+    public List<Node> getNodeList()
     {
-        State state = new State(nodeRepository.calculatePathDistance());
-
-        while (temperature > minTemperature)
-        {
-            nodeRepository.shuffle();
-            State nextState = new State(nodeRepository.calculatePathDistance());
-
-            System.out.println("Current energy " + state.getEnergy());
-            System.out.println("next state energy " + nextState.getEnergy());
-
-            double deltaEnergy = state.getEnergy() - nextState.getEnergy();
-
-            System.out.println("Delta energy " + deltaEnergy);
-
-            if (deltaEnergy >= 0)
-            {
-                state = nextState;
-            }
-            else if (getProbability(deltaEnergy) < Math.random())
-            {
-                state = nextState;
-                System.out.println("Changed to higher energy state with probability of " + getProbability(deltaEnergy) + " where temp was " + temperature);
-            }
-
-            System.out.println("New Current energy " + state.getEnergy());
-            System.out.println();
-
-            decreaseTemperature();
-        }
-
-        return state;
+        return nodeRepository.findAll();
     }
 
-    private void decreaseTemperature()
+    public void insertNodes(List<Node> nodeList)
     {
-        temperature -= temperatureConst;
-    }
-
-    private double getProbability(double deltaEnergy)
-    {
-        return Math.exp(-deltaEnergy / temperature);
+        nodeRepository.reassignRepository(nodeList);
     }
 }
