+++ src/main/resources/application.properties
@@ -1,7 +1,1 @@
-logging.level.org.springframework.jdbc=DEBUG
-#logging.level.com.zaxxer.hikari.HikariConfig=DEBUG
 
-# or use devtools
-spring.h2.console.enabled=true
-spring.datasource.generate-unique-name=false
-spring.datasource.name=node_repository0000000000000000000000
